refactor(nms): Consolidate PolicyContext and fix GatewayPoolsContext issues
